// Zybach.Web/src/app/shared/models/report-template-update-dto.ts
/**
 * Zybach
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v1
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export class ReportTemplateUpdateDto {
    FileResource?: Blob;
    DisplayName: string;
    Description?: string;
    ReportTemplateModelID: number;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}